Create logic for cleaning up phone numbers
The logic takes the phone number and filters it down to just the numbers
leaving out any other symbols such as hyphens and parentheses. It then
returns a number if it is 10 characters long, or 11 character long if it
starts with a 1, and all other numbers are tossed out as invalid.